refactor(Card): Move Card component to SidebarTree as TreeInfos
